added the ability to have arrows back on the stream lines
Former-commit-id: 967fd1d367c1a232b7ce015fe7a688c8812e874f

+++ src/ves/xplorer/event/viz/cfdStreamers.h
@@ -104,7 +104,7 @@ private:
     float propagationTime;
     float integrationStepLength;
     float lineDiameter;
-    float arrowDiameter;
+    double arrowDiameter;
 
     unsigned int xValue;///>number of points for x direction
     unsigned int yValue;///>number of points for y direction
